+++ client/src/components/characterCustomization/Borders.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import Grid from '@mui/material/Unstable_Grid2';
-import Tick from '../assets/images/greenTick.png';
+import Tick from '../../assets/images/greenTick.png';
 
-const images = require.context('../assets/images/borders', false, /\.(png|gif)$/);
+const images = require.context('../../assets/images/borders', false, /\.(png|gif)$/);
 const imageList = images.keys().map(image => images(image));
 var selected = '0';
